Implement CRL and AIA extension (#35)
- Add CRL Distribution Endpoint extension to Issuer CA certs
- Add Authority Information Access extension to subscriber certs
- Add endpoints to service
- UI bugfixes and test improvements

+++ src/KeyVault/KeyVaultServiceClient.cs
@@ -13,14 +13,15 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Azure.IIoT.Diagnostics;
+using Microsoft.Azure.IIoT.Exceptions;
 using Microsoft.Azure.KeyVault;
 using Microsoft.Azure.KeyVault.Models;
 using Microsoft.Azure.KeyVault.WebKey;
 using Microsoft.Azure.Services.AppAuthentication;
 using Microsoft.IdentityModel.Clients.ActiveDirectory;
 using Microsoft.Rest;
+using Microsoft.Rest.Azure;
 using Opc.Ua;
-using static Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault.KeyVaultCertFactory;
 
 namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
 {
@@ -29,8 +30,9 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
         public X509Certificate2 Certificate { get; set; }
         public string KeyIdentifier { get; set; }
     }
+
     /// <summary>
-    ///
+    /// The KeyVault service client.
     /// </summary>
     public class KeyVaultServiceClient
     {
@@ -48,17 +50,21 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
         public const string TagIssuerList = "Issuer";
         public const string TagTrustedList = "Trusted";
 
-        // TODO: make group secret name configurable
-        public const string GroupSecret = "groups";
-
         /// <summary>
-        ///
+        /// Create the service client for KeyVault, with user or service credentials.
         /// </summary>
+        /// <param name="groupSecret">The name of the secret for group configuration</param>
         /// <param name="vaultBaseUrl">The Url of the Key Vault.</param>
         /// <param name="keyStoreHSM">The KeyVault is HSM backed.</param>
         /// <param name="logger">The logger.</param>
-        public KeyVaultServiceClient(string vaultBaseUrl, bool keyStoreHSM, ILogger logger)
+        public KeyVaultServiceClient(
+            string groupSecret,
+            string vaultBaseUrl,
+            bool keyStoreHSM,
+            ILogger logger
+            )
         {
+            _groupSecret = groupSecret;
             _vaultBaseUrl = vaultBaseUrl;
             _keyStoreHSM = keyStoreHSM;
             _logger = logger;
@@ -136,7 +142,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
         /// </summary>
         public async Task<string> GetCertificateConfigurationGroupsAsync(CancellationToken ct = default)
         {
-            SecretBundle secret = await _keyVaultClient.GetSecretAsync(_vaultBaseUrl, GroupSecret, ct).ConfigureAwait(false);
+            SecretBundle secret = await _keyVaultClient.GetSecretAsync(_vaultBaseUrl, _groupSecret, ct).ConfigureAwait(false);
             return secret.Value;
         }
 
@@ -145,7 +151,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
         /// </summary>
         public async Task<string> PutCertificateConfigurationGroupsAsync(string json, CancellationToken ct = default)
         {
-            SecretBundle secret = await _keyVaultClient.SetSecretAsync(_vaultBaseUrl, GroupSecret, json, null, ContentTypeJson, null, ct).ConfigureAwait(false);
+            SecretBundle secret = await _keyVaultClient.SetSecretAsync(_vaultBaseUrl, _groupSecret, json, null, ContentTypeJson, null, ct).ConfigureAwait(false);
             return secret.Value;
         }
 
@@ -164,29 +170,55 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
         /// Read all certificate versions of a CA certificate group.
         /// </summary>
         /// <param name="id"></param>
+        /// <param name="thumbprint">filter for thumbprint</param>
+        /// <param name="nextPageLink"></param>
+        /// <param name="pageSize"></param>
         /// <param name="ct"></param>
         /// <returns></returns>
-        public async Task<X509Certificate2Collection> GetCertificateVersionsAsync(string id, CancellationToken ct = default)
+        public async Task<(X509Certificate2Collection, string)> GetCertificateVersionsAsync(string id, string thumbprint = null, string nextPageLink = null, int? pageSize = null, CancellationToken ct = default)
         {
             var certificates = new X509Certificate2Collection();
+            pageSize = pageSize ?? MaxResults;
             try
             {
-                var certItems = await _keyVaultClient.GetCertificateVersionsAsync(_vaultBaseUrl, id, MaxResults, ct).ConfigureAwait(false);
+                IPage<CertificateItem> certItems = null;
+                if (nextPageLink != null)
+                {
+                    certItems = await _keyVaultClient.GetCertificateVersionsNextAsync(nextPageLink, ct).ConfigureAwait(false);
+                }
+                else
+                {
+                    certItems = await _keyVaultClient.GetCertificateVersionsAsync(_vaultBaseUrl, id, pageSize, ct).ConfigureAwait(false);
+                }
                 while (certItems != null)
                 {
                     foreach (var certItem in certItems)
                     {
-                        var certBundle = await _keyVaultClient.GetCertificateAsync(certItem.Id, ct).ConfigureAwait(false);
-                        var cert = new X509Certificate2(certBundle.Cer);
-                        certificates.Add(cert);
+                        if (certItem.Attributes.Enabled ?? false)
+                        {
+                            var certBundle = await _keyVaultClient.GetCertificateAsync(certItem.Id, ct).ConfigureAwait(false);
+                            var cert = new X509Certificate2(certBundle.Cer);
+                            if (thumbprint == null ||
+                                cert.Thumbprint.Equals(thumbprint, StringComparison.OrdinalIgnoreCase))
+                            {
+                                certificates.Add(cert);
+                            }
+                        }
                     }
                     if (certItems.NextPageLink != null)
                     {
-                        certItems = await _keyVaultClient.GetCertificateVersionsNextAsync(certItems.NextPageLink, ct).ConfigureAwait(false);
+                        nextPageLink = certItems.NextPageLink;
+                        certItems = null;
+                        if (certificates.Count < pageSize)
+                        {
+                            certItems = await _keyVaultClient.GetCertificateVersionsNextAsync(nextPageLink, ct).ConfigureAwait(false);
+                            nextPageLink = null;
+                        }
                     }
                     else
                     {
                         certItems = null;
+                        nextPageLink = null;
                     }
                 }
             }
@@ -194,7 +226,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
             {
                 _logger.Error("Error while loading the certificate versions for " + id + ".", () => new { ex });
             }
-            return certificates;
+            return (certificates, nextPageLink);
         }
 
         /// <summary>
@@ -374,6 +406,8 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
                 // intentionally ignore errors
             }
 
+            string caTempCertIdentifier = null;
+
             try
             {
                 // policy self signed, new key
@@ -399,7 +433,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
                 }
                 var createdCertificateBundle = await _keyVaultClient.GetCertificateAsync(_vaultBaseUrl, id).ConfigureAwait(false);
                 var caCertKeyIdentifier = createdCertificateBundle.KeyIdentifier.Identifier;
-                var caTempCertIdentifier = createdCertificateBundle.CertificateIdentifier.Identifier;
+                caTempCertIdentifier = createdCertificateBundle.CertificateIdentifier.Identifier;
 
                 // policy unknown issuer, reuse key
                 var policyUnknownReuse = CreateCertificatePolicy(subject, keySize, false, true);
@@ -455,18 +489,31 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
                     new X509Certificate2Collection(signedcert)
                     );
 
-                // disable the temp cert for self signing operation
-                var attr = new CertificateAttributes()
-                {
-                    Enabled = false
-                };
-                await _keyVaultClient.UpdateCertificateAsync(caTempCertIdentifier, null, attr);
-
                 return signedcert;
             }
-            catch
+            catch (KeyVaultErrorException kex)
             {
+                var ex = kex;
                 throw new ServiceResultException(StatusCodes.BadInternalError, "Failed to create new Root CA certificate");
+            }
+            finally
+            {
+                if (caTempCertIdentifier != null)
+                {
+                    try
+                    {
+                        // disable the temp cert for self signing operation
+                        var attr = new CertificateAttributes()
+                        {
+                            Enabled = false
+                        };
+                        await _keyVaultClient.UpdateCertificateAsync(caTempCertIdentifier, null, attr);
+                    }
+                    catch
+                    {
+                        // intentionally ignore error
+                    }
+                }
             }
         }
 
@@ -490,6 +537,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
             int keySize,
             int hashSize,
             KeyVaultSignatureGenerator generator,
+            string authorityInformationAccess,
             CancellationToken ct = default)
         {
             CertificateOperation createResult = null;
@@ -538,7 +586,8 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
                     (ushort)hashSize,
                     issuerCert,
                     publicKey,
-                    generator);
+                    generator,
+                    extensionUrl: authorityInformationAccess);
 
                 // merge signed cert with keystore
                 var mergeResult = await _keyVaultClient.MergeCertificateAsync(
@@ -590,7 +639,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
         {
             try
             {
-                string secretIdentifier = CrlSecretName(id, certificate);
+                string secretIdentifier = CrlSecretName(id, certificate.Thumbprint);
                 SecretAttributes secretAttributes = new SecretAttributes()
                 {
                     Enabled = true,
@@ -620,7 +669,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
         /// </summary>
         public async Task<Opc.Ua.X509CRL> LoadIssuerCACrl(string id, X509Certificate2 certificate, CancellationToken ct = default)
         {
-            string secretIdentifier = CrlSecretName(id, certificate);
+            string secretIdentifier = CrlSecretName(id, certificate.Thumbprint);
             var secret = await _keyVaultClient.GetSecretAsync(_vaultBaseUrl, secretIdentifier, ct).ConfigureAwait(false);
             if (secret.ContentType == ContentTypeCrl)
             {
@@ -629,6 +678,29 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
             }
             return null;
         }
+
+        /// <summary>
+        /// Load CRL by ThumbPrint in group.
+        /// </summary>
+        public async Task<Opc.Ua.X509CRL> LoadIssuerCACrl(string id, string thumbPrint, CancellationToken ct = default)
+        {
+            try
+            {
+                string secretIdentifier = CrlSecretName(id, thumbPrint);
+                var secret = await _keyVaultClient.GetSecretAsync(_vaultBaseUrl, secretIdentifier, ct).ConfigureAwait(false);
+                if (secret.ContentType == ContentTypeCrl)
+                {
+                    var crlBlob = Convert.FromBase64String(secret.Value);
+                    return new Opc.Ua.X509CRL(crlBlob);
+                }
+            }
+            catch (KeyVaultErrorException)
+            {
+                // hide KeyVault Exception
+            }
+            throw new ResourceNotFoundException("A CRL for this thumbprint doesn't exist.");
+        }
+
 
         /// <summary>
         /// Imports a Private Key for group id and certificate.
@@ -849,13 +921,15 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
         /// <summary>
         /// Purge all CRL and Certificates groups. Use for unit test only!
         /// </summary>
-        public async Task PurgeAsync(CancellationToken ct = default)
+        public async Task PurgeAsync(string configId = null, string groupId = null, CancellationToken ct = default)
         {
             var secretItems = await _keyVaultClient.GetSecretsAsync(_vaultBaseUrl, MaxResults, ct).ConfigureAwait(false);
             while (secretItems != null)
             {
                 foreach (var secretItem in secretItems.Where(s =>
-                    (s.ContentType == ContentTypeCrl || s.ContentType == ContentTypePem || s.ContentType == ContentTypePfx)
+                    ((s.ContentType == ContentTypeCrl || s.ContentType == ContentTypePem || s.ContentType == ContentTypePfx) &&
+                    (groupId == null || s.Identifier.Name.StartsWith(groupId, StringComparison.OrdinalIgnoreCase))) ||
+                    s.Identifier.Name.Equals(configId, StringComparison.OrdinalIgnoreCase)
                     ))
                 {
                     try
@@ -884,16 +958,19 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
             {
                 foreach (var certItem in certItems)
                 {
-                    try
+                    if (groupId == null ||
+                        groupId.Equals(certItem.Identifier.Name, StringComparison.OrdinalIgnoreCase))
                     {
-                        var deletedCertBundle = await _keyVaultClient.DeleteCertificateAsync(_vaultBaseUrl, certItem.Identifier.Name, ct);
-                        await _keyVaultClient.PurgeDeletedCertificateAsync(_vaultBaseUrl, certItem.Identifier.Name, ct);
+                        try
+                        {
+                            var deletedCertBundle = await _keyVaultClient.DeleteCertificateAsync(_vaultBaseUrl, certItem.Identifier.Name, ct);
+                            await _keyVaultClient.PurgeDeletedCertificateAsync(_vaultBaseUrl, certItem.Identifier.Name, ct);
+                        }
+                        catch
+                        {
+                            // intentionally fall through, purge may fail
+                        }
                     }
-                    catch
-                    {
-                        // intentionally fall through, purge may fail
-                    }
-
                 }
                 if (certItems.NextPageLink != null)
                 {
@@ -991,9 +1068,9 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
         {
             return id + "Key" + requestId;
         }
-        private string CrlSecretName(string id, X509Certificate2 certificate)
+        private string CrlSecretName(string id, string thumbprint)
         {
-            return id + "Crl" + certificate.Thumbprint;
+            return id + "Crl" + thumbprint;
         }
 
         private async Task<X509CRL> LoadCrlSecret(string secretIdentifier, CancellationToken ct = default)
@@ -1031,6 +1108,7 @@ namespace Microsoft.Azure.IIoT.OpcUa.Services.Vault.KeyVault
             throw new Exception("Unknown Private Key format.");
         }
 
+        private readonly string _groupSecret;
         private readonly string _vaultBaseUrl;
         private readonly bool _keyStoreHSM;
         private IKeyVaultClient _keyVaultClient;
